added function to gather custom facts
The Get-CustomFacts function will use the facts_path variable on Windows to parse that directory for ps1 files. Each file will be executed and is expected to return a value which will be turned into a custom fact. The fact itself will be named after the filename (without extension) that was executed.